Consolidate tests into a single root test/ directory
- Move API tests to test/api/, DB scripts to test/db/, Playwright suite to test/e2e/
- Move verify-app.mjs orchestrator to test/
- Update all path references (playwright config, helpers, verify-app, db scripts)
- Update package.json scripts, .gitignore, and README with new layout
