+++ .gitignore
@@ -11,8 +11,8 @@
 /node_modules
 
 # e2e artifacts
-/e2e/test-results
-/e2e/playwright-report
+/test/e2e/test-results
+/test/e2e/playwright-report
 /test-results
 /playwright-report
 /.last-run.json
