+++ README.md
@@ -96,10 +96,22 @@ npm run test:front   # frontend unit tests
 npm run test:build   # frontend production build
 ```
 
+## Test layout
+
+All tests and test tooling live under the root `test/` directory:
+
+```
+test/
+  api/            # API integration tests (node --test style, run against the backend)
+  db/             # database setup/reset scripts used by the API and E2E tests
+  e2e/            # Playwright GUI tests + shared helpers
+  verify-app.mjs  # full-stack verification orchestrator (npm run verify)
+```
+
 ## End-to-end (GUI) tests with Playwright
 
-The `e2e/` suite drives the real UI in a headless browser and covers the main
-user flows: login (success and failure), account listing, search, and
+The `test/e2e/` suite drives the real UI in a headless browser and covers the
+main user flows: login (success and failure), account listing, search, and
 create/edit/delete with server-side persistence checks.
 
 Install the browser once (already available in the Cursor cloud agent
@@ -112,14 +124,14 @@ npm run test:e2e
 
 Playwright automatically:
 
-1. resets and seeds the MariaDB 10.11 test database (`scripts/reset-test-db.sh`),
+1. resets and seeds the MariaDB 10.11 test database (`test/db/reset-test-db.sh`),
 2. starts the backend (`back/`) and the Vite dev server (`front/`),
-3. runs the specs in `e2e/tests/`.
+3. runs the specs in `test/e2e/`.
 
 Useful options:
 
 ```bash
-npm run test:e2e -- e2e/tests/crud.spec.js   # run a single spec
+npm run test:e2e -- test/e2e/crud.spec.js    # run a single spec
 PW_VIDEO=on PW_SLOWMO=350 npm run test:e2e    # record videos in slow motion
 npm run test:e2e:report                       # open the HTML report
 ```
